scrapers/requirements.txt
requests==2.22.0
beautifulsoup4==4.7.1
psycopg2-binary==2.8.3
